+++ Changelog.md
@@ -109,3 +109,4 @@
     * Removed "Waiting" table
     * Added new scene for awaited series
     * Fixed problem with series that are not complete and do not provide a continuous list of episodes
+    * Added option to delete a series via right-click